Move abstract definitions to own namespace and document them
