+++ app/src/main/java/dev/patrickgold/florisboard/ime/core/InputEventDispatcher.kt
@@ -19,7 +19,7 @@ package dev.patrickgold.florisboard.ime.core
 import android.os.SystemClock
 import dev.patrickgold.florisboard.BuildConfig
 import dev.patrickgold.florisboard.ime.text.key.KeyCode
-import dev.patrickgold.florisboard.ime.text.key.TextKeyData
+import dev.patrickgold.florisboard.ime.text.keyboard.TextKeyData
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.Channel
 import timber.log.Timber
