+++ app/src/main/java/dev/patrickgold/florisboard/ime/keyboard/Keyboard.kt
@@ -1,0 +1,49 @@
+/*
+ * Copyright (C) 2021 Patrick Goldinger
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.patrickgold.florisboard.ime.keyboard
+
+/**
+ * Abstract class describing a computed keyboard. The exact implementation is dependent on the subclass and the
+ * structure can vary quite much between different subclasses.
+ */
+abstract class Keyboard {
+    /**
+     * Returns the key for given [pointerX] and [pointerY] coords or null if no key touch hit box is defined at the
+     * given coords.
+     *
+     * @param pointerX The x-coordinate of the input event, absolute within the parent keyboard view.
+     * @param pointerY The y-coordinate of the input event, absolute within the parent keyboard view.
+     *
+     * @return The key for given coords or null if no key touch hit box is defined for this position.
+     */
+    abstract fun getKeyForPos(pointerX: Int, pointerY: Int): Key?
+
+    /**
+     * Returns an iterator which allows to loop through all keys within the layout, independent of the actual
+     * structure and layout.
+     */
+    abstract fun keys(): Iterator<Key>
+
+    /**
+     * Layouts the keys according the the dimensions and parameters provided by given [keyboardView]. This method's
+     * exact behavior is highly dependent aon the actual subclass.
+     *
+     * @param keyboardView The parent keyboard view which is used to gather the absolute dimensions and other
+     *  subclass specific parameters.
+     */
+    abstract fun layout(keyboardView: KeyboardView)
+}
