+++ app/src/main/java/dev/patrickgold/florisboard/ime/popup/PopupSet.kt
@@ -16,7 +16,7 @@
 
 package dev.patrickgold.florisboard.ime.popup
 
-import dev.patrickgold.florisboard.ime.text.key.KeyData
+import dev.patrickgold.florisboard.ime.keyboard.KeyData
 import dev.patrickgold.florisboard.ime.text.key.KeyHintMode
 import dev.patrickgold.florisboard.ime.text.keyboard.TextComputingEvaluator
 import kotlinx.serialization.Serializable
