+++ app/src/main/java/dev/patrickgold/florisboard/ime/text/gestures/GlideTypingManager.kt
@@ -6,8 +6,8 @@ import dev.patrickgold.florisboard.ime.core.Subtype
 import dev.patrickgold.florisboard.ime.extension.AssetManager
 import dev.patrickgold.florisboard.ime.extension.AssetRef
 import dev.patrickgold.florisboard.ime.extension.AssetSource
+import dev.patrickgold.florisboard.ime.keyboard.Key
 import dev.patrickgold.florisboard.ime.text.TextInputManager
-import dev.patrickgold.florisboard.ime.text.keyboard.Key
 import kotlinx.coroutines.*
 import org.json.JSONObject
